Fix Vimeo embed url
The previous url only seemed to work on a local browser, but not when pushed
live. This new url was copied from vimeo.com and is thus slightly updated.
There are a few parameters missing still, however. Below is a sample of what a
"proper" embed should be. We are still missing the fullscreen options.

  <iframe src="//player.vimeo.com/video/8567074" width="500" height="281" webkitallowfullscreen mozallowfullscreen allowfullscreen>
  </iframe>