Geometry Nodes: Color Ramp
This is using the node from compositor. I need to talk to Aaron about that
because the examples being also shared may be confusing to users.

That is even worse for things like the combine nodes which are all in a single
page for the other systems, but for geometry nodes they are in different
categories.