Geometry Node: Combine RGB
I deliberately not re-used the compositor node page. That is a single
page with all combine/separate. Although handy, it does not follow the
categories and granularity of the geometry nodes.

For that reason, for now, I kept the page simple and to the point.

Also there is no bpy.types defined yet.