Fix latest manual version warning that it is not the latest version
This problem was introduced in r6839, assuming the dev version is the latest
official release is not correct.

The mechanism for this warning is still broken and it's not clear that it ever
worked well. For that reason it is disabled. See code comments for what needs
to be fixed.



git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk/blender_docs@7429 c4de1f47-6596-e411-a384-0024e86c2797