Update Pip install command
Sphinx is a bit slow with their update policy on 3rd party libraries,
currently they target packages that are bundled with Ubuntu 16.04
package manager which is a bit old when we are using the latest version of sphinx.
This also causes some issues where environments can vary depending when sphinx was first installed.

git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk/blender_docs@7811 c4de1f47-6596-e411-a384-0024e86c2797